Fix GLIBC errors seen by some
The crux of this issue is that the version of glibc on the github
machine that builds this vs that of those who run it may be different.
We need to use this env variables build in the libraries instead of
dynamically loading them on the host machine at runtime.