Job address now being sent with scheduled jobs
We want to allow contracts to verify that they were called by a job
with the desired permissions to run the contract.  To do this, we need
to pass in an identifier of the job that is being run.

We were going to pass in the public key of the contract, but after some
research, it seems this is not retrievable here.  Instead, we are
passing the address of the contract.