Better message limit handling (#915)
* Don't always require a limit on messages

Adding an option to get all messages, not just N messages.  This is
useful for times when we are going to filter the messages downstream.

* Move message response counts

If we request a number of messages, we want that to be the number
returned, not the number before filtering

Also fixing places that wanted all messages.  Now that we can return
all, let's just request all

* Return all messages for requested messages in queue

* Linting

---------

Co-authored-by: Tyler Ruppert <{ID}+{username}@users.noreply.github.com>