Add cli command to delete any stuck smart contract deployments
Paloma tracks smart contract deployments purely as a caching mechanism
to prevent double-deploying too quickly.  We have an edge case where a
deployment fails to go through for too long, it will get stuck.  Paloma
will think it's deploying, but the message was long since deleted.

Now, we have a way of deleting this stuck deployment.  This is
effectively the same as clearing the cache flag, and the chain will
trigger a fresh deployment on the next block.