feat: Migrate Pigeon TTL to block height
The previous Pigeon keep alive implementation would use a five minute
TTL and compare against the current block time. This would frequently
lead to problems after a chain halt, during which Pigeons would not
be able to refresh their TTL.

This change removes the time based approach and substitutes the
block time with block height, with a default of 185 blocks of
lifetime before a Pigeon is considered stale.