feat: Add jail sentence backoff
This change introduces a dynamic backoff that increases the given
sentence to a jailed validator should they become jailed repeatedly
within short periods of time.

This is done to incentivise validators to examine their setup
and ensure they remain part of the active validator set, as well
as reduce the amount of valset changes resulting from frequent
jailings.