chore: reduce consensus msg lifetime
Changes the lifetime of a consensus message to 300 blocks,
approximately 8 minutes with an average block speed of 1.6
seconds per block.

Before automatic reassignment of messages, a message would
stay in the queue for up to 1200 blocks, giving the assigned
validator some time to fix ongoing issues.

With the reassignment of tardy messages in place, there's
no need for messages to stay around considerably longer
than 100 blocks really, but with 300 blocks we still get
some opportunity to inspect what's going on in logs.

We can reduce this value further to something like 100 or
even 50 once we're confident with our relay stability.