feat: add support for ERC20 relinking during  contract deployment
This change introduces an extended workflow when running
contract upgrade. Very simply put, the way it works is by

- while attesting to an uploaded contract, inspect deployed tokens on
  target chain
- if no tokens deployed, immediately active the new contract
- if 1 or more tokens deployed, send messages to the old contract to
  start executing the relink
- keep track of the message IDs used for this, finish the attestation
  but do not enable the new contract yet

- while attesting to a SLC message, verify that the message ID is
  part of an ongoing erC20 ownership transfer
- if so, check if the message failed
- if so, attempt to retry, update the deployment setting this message ID
  to failed and adding the new message ID for the retry
- if the message succeeded, update the deployment by setting the message
  ID to successfully processed.
- next, if more erc20 tokens need relinking for this deployment, do
  nothing
- otherwise, mark the deployment as successful and activate the new
  chain

This new flow adds a large amount of complexity due to the asynchronous
and distributed workflow that's now governing a successful contract
deployment. Code quality and test coverage is nowhere near where I'd
like to see it, but I need to move on to Pigeon Feed for now.

I'd like to revisit this topic in the future to move this entire topic
into its own module that's responsible for contract deployment
maintenance and monitoring.