feat: paloma bridge exit tax (#1196)
# Related Github tickets

- https://github.com/VolumeFi/paloma/issues/1709

# Background

We're going to charge a tax fee for every transfer out of paloma. The
tax rate is defined by governance vote and we can set exceptions to
tokens and addresses.

- Add new governance vote to gravity module
- The bridge tax information is added to genesis
- Before adding transfers to the batched pool in gravity, we set the tax
amount, if applicable. This amount is removed from the total amount to
transfer. Compass only sees the decreased amount.
- On receiving the transfer complete event we burn the total amount of
tokens (transferred + taxed)
- On cancelling a transfer, we return the total amount including tax

# Testing completed

- [x] test coverage exists or has been added/updated
- [ ] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.