fix: add gravity proposals so they can be decoded
We already have gov proposals from the gravity module on chain, so we
need to register them in order for them to be decoded.

We need to add them in the `gravity` proto directory, so the proto file
names match the proto package, otherwise we get warnings on the command
line.

The resulting gravity types reside within the `skyway/types` dir.