feat: add support for gas estimation on CQ messages (#1238)
# Related Github tickets

- https://github.com/VolumeFi/paloma/issues/344
- https://github.com/VolumeFi/paloma/issues/1761

# Background

> [!IMPORTANT]  
> This will only merged back to master after Pigeon changes and
successful smoke tests on PTN

This change adds the infrastructure for Paloma to support gas
estimation. Most notably, it

- changes the lifecycle of the consensus queue message
- adds endpoints for Pigeons to query for and report gas estimation
- uses an endblocker to process any estimated messages

## Lifecycle changes

The current message lifecycle is very bare bone. Pigeons ask for
messages to be signed, and ultimately relay it. There is little
filtering or state processing done. A message will even be handed out
for relaying long before it has reached consensus on signatures (Pigeon
checks that, not Paloma).

This change adds the concept of gas estimation to all consensus queue
messages. It's not super pretty, since really only EVM messages will
need the data. However, the CQ module is the main interface for managing
the queue and message lifecycles. It's already grossly interwoven with
the EVM module, and the whole idea of abstraction has not really paid
off. We hope to address this in the future.

Paloma will hand out messages for gas estimation once they have reached
consensus on signatures. Pigeons will not relay a message that requires
gas estimation and has not yet received one. Even though a message might
have been fully signed, it won't be relayed until gas estimation is
completed.

For this, Pigeons will query the message, build the payload and estimate
the gas using their configured RPC provider. They will send their
estimate to Paloma, and Paloma will start tracking every estimate. We
expect estimates to be different across RPC providers, and have settled
to use the median value of all collected estimates, with a generous
`1.2` multiplier for safety margins.

Paloma will continuously monitor messages and check whether they have
received enough estimates to calculate the estimate. IF the message
happens to be an SLC message, Paloma will then proceed to automatically
calculated the fees based on the relayer fee settings for this message.

Once a message has been estimated, all signatures are removed from the
message and the signing process starts again, this time using the
embedded information (fees & estimates). After this second round of
signatures, the messages will finally be sent.

# Testing completed

- [x] test coverage exists or has been added/updated
- [ ] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.

---------

Co-authored-by: Luis Carvalho <mail.lmcarvalho@gmail.com>