feat: add fee manager to consensus (#1245)
# Background

The fee manager smart contract will be independent from compass and the
bridge. It will need to be deployed manually before compass, and
ownership will have to be passed to compass after its deployment.

> [!CAUTION]
> There is no support yet for automatic compass upgrades, as they will
require a handover of the fee manager just like ERC20 tokens. This will
be added in a follow up MR.

# Testing completed

- [x] test coverage exists or has been added/updated
- [ ] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.