feat: add selective gas exemption whitelist (#1332)
# Related Github tickets

- https://github.com/VolumeFi/paloma/issues/2487

# Background

This change introduces a new governance concept: a whitelist of gas
exempt addresses. The list can be updated by creating a
`MsgUpdateParams` vote on the `paloma` module.

Every member address of the list will be exempt from paying gas fees on
the chain.

The change also brings some QOL improvements in the integration testing
package.

# Testing completed

- [x] test coverage exists or has been added/updated
- [x] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.